+++ tests/source/csr_test.cpp
@@ -18,7 +18,8 @@ TEST(csr_test, basic_washington_test_generates_and_converts_to_csr) {
 }
 
 TEST(csr_test, random_washington_test_generates_and_converts_to_csr) {
-    int n = std::uniform_int_distribution<int>(1, 1e5)(generator);
+    int n = std::uniform_int_distribution<int>(1,
+            static_cast<int>(1e5))(generator);
     ASSERT_NO_THROW(CSR csr = washington_test(n, true));
 }
 
@@ -27,7 +28,8 @@ TEST(csr_test, basic_zadeh_test_generates_and_converts_to_csr) {
 }
 
 TEST(csr_test, random_zadeh_test_generates_and_converts_to_csr) {
-    int n = std::uniform_int_distribution<int>(1, 1e3)(generator);
+    int n = std::uniform_int_distribution<int>(1,
+            static_cast<int>(1e3))(generator);
     ASSERT_NO_THROW(CSR csr = zadeh_test(n, true));
 }
 
@@ -60,7 +62,8 @@ TEST(al_test, run_Dijkstra_on_not_weighted_basic_washington_test) {
 }
 
 TEST(csr_test, run_dfs_on_random_washington_test) {
-    int n = std::uniform_int_distribution<int>(1, 1e3)(generator);
+    int n = std::uniform_int_distribution<int>(1,
+            static_cast<int>(1e3))(generator);
     CSR csr = washington_test(n, true);
     std::vector <int> expected_dfs_result;
     expected_dfs_result.push_back(1);
